forge -  restructured memstore
